Move toPem into OpenSSL class
