+++ systemtests/src/main/java/io/enmasse/systemtest/certs/ssl/CertPair.java
@@ -3,7 +3,7 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-package io.enmasse.systemtest.certs.openssl;
+package io.enmasse.systemtest.certs.ssl;
 
 import java.io.Closeable;
 import java.io.File;
