+++ systemtests/src/main/java/io/enmasse/systemtest/certs/ssl/CertSigningRequest.java
@@ -2,7 +2,7 @@
  * Copyright 2017-2018, EnMasse authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.enmasse.systemtest.certs.openssl;
+package io.enmasse.systemtest.certs.ssl;
 
 import java.io.Closeable;
 import java.io.File;
